% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miami_input.R
\name{check_miami_input}
\alias{check_miami_input}
\title{Check input to miami plot}
\usage{
check_miami_input(data, split_by, split_at, chr = "chr", pos = "pos", p = "p")
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{split_by}{A character vector. The name of the column to use for
splitting into upper and lower sections of the Miami plot. Required.}

\item{split_at}{A character or numeric vector. If numeric, the upper plot
will contain your results where the values in the \code{split_by} column
are >= \code{split_at}. If character, upper plot will contain your results
where the values in the \code{split_by} column are equal to
\code{split_at}. Required.}

\item{chr}{The name of the column containing your chromosome information.
Defaults to "chr"}

\item{pos}{The name of the column containing your position information.
Defaults to "pos"}

\item{p}{The name of the column containing your p-value information.
Defaults to "p"}
}
\value{
If no problems, executes silently. Will return an error message if
  there are problems.
}
\description{
Check input to miami plot
}
\examples{
check_miami_input(data = gwas_results, split_by = "beta",
                            split_at = 0)
}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
