% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_col_info.R
\name{find_col_info}
\alias{find_col_info}
\title{Find column index matching a given string and verify numeric or integer class.}
\usage{
find_col_info(data, col)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{col}{An object name or character vector. The name of the column to
search the data.frame for and check class. Required.}
}
\value{
Will return an error message if there are problems. Otherwise returns
  the index of the named column in the dataframe.
}
\description{
Find column index matching a given string and verify numeric or integer class.
}
\examples{
If you are searching for the "chr" column.
find_col_info(data = df, col = "chr")

}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
