% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_miami.R
\name{highlight_miami}
\alias{highlight_miami}
\title{Make highlight data.frame for miami plot}
\usage{
highlight_miami(
  data,
  highlight,
  highlight_col,
  highlight_color = "green",
  logged_p = "logged_p",
  rel_pos = "rel_pos"
)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{highlight}{A vector of SNPs or gene names you would like to highlight.
Defaults to NULL. If you specify this, you must also specify highlight_col
so that we know where to find the items.}

\item{highlight_col}{The column where the values you provide in
\code{highlight} can be found.}

\item{highlight_color}{The color that you would like the highlighted SNPs or
genes to be. Defaults to "green."}

\item{logged_p}{The name of the column containing your logged p-value
information. Defaults to "logged_p," assuming data preparation with
\code{prep_miami_data}}

\item{rel_pos}{The name of the column containing the positon of each
SNP/probe relative to all other SNPs/probes in the genome. Defaults to
"rel_pos," assuming data preparation with \code{prep_miami_data}.}
}
\value{
A data.frame with three columns: relative position, logged p-value,
  and highlight color.
}
\description{
Make highlight data.frame for miami plot
}
\examples{
 \dontrun{
 # You can provide a list of SNP / gene names
 upper_highlight <- highlight_miami(data = plot_data$upper,
                                     highlight = c(rs1, rs2, rs3, ..., rsN),
                                     highlight_col = "rsid")

 # Or an object containing those names.
 upper_highlight <- highlight_miami(data = plot_data$upper,
                                     highlight = snps_of_interest,
                                     highlight_col = "rsid")

 # When highlighting, the items in 'highlight' must all come from a single
 # column. Specifying multiple columns in highlight_col will return an error.
 highlight_miami(data = df, split_by = "study", split_at = "A",
                 highlight = snps_of_interest,
                 highlight_col = c("rsid", "gene"))
 }

}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
