% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_miami_labels.R
\name{make_miami_labels}
\alias{make_miami_labels}
\title{Make label data.frame for miami plot}
\usage{
make_miami_labels(
  data,
  hits_label_col,
  hits_label = NULL,
  top_n_hits = 5,
  loggedp = "loggedp",
  rel_pos = "rel_pos"
)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{hits_label_col}{Either the name of the column(s), max = 2, to use for
automatically labeling n hits, determined using \code{top_n_hits}, or the
column where the values you provide in \code{hits_label} can be found.
Required.}

\item{hits_label}{A user-specified character vector of probes/genes/SNPs
to label. Defaults to NULL.}

\item{top_n_hits}{How many of the top hits do you want to label? Defaults to
5. Set to NULL to turn off this filtering, but this is not recommended
because the plot can get cluttered easily.}

\item{loggedp}{The name of the column containing your logged p-value
information. Defaults to "loggedp," assuming data preparation with
\code{prep_miami_data}}

\item{rel_pos}{The name of the column containing the positon of each
SNP/probe relative to all other SNPs/probes in the genome. Defaults to
"rel_pos," assuming data preparation with \code{prep_miami_data}.}
}
\value{
A data.frame with three columns: relative position, logged p-value,
  and the label to add to the miami plot. Designed to be used with ggrepel.
}
\description{
Make label data.frame for miami plot
}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
