% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamiplot.R
\docType{package}
\name{miamiplot-package}
\alias{miamiplot}
\alias{miamiplot-package}
\title{miamiplot: Create a ggplot2 miami plot}
\description{
This package can be used to create a Miami plot in the form of a 
    ggplot2 object. Users can choose significance lines, colors, and labels.
}
\author{
\strong{Maintainer}: Julie White \email{julie.white@nih.gov}

}
\keyword{internal}
