% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_miami_data.R
\name{prep_miami_data}
\alias{prep_miami_data}
\title{Prepare data for miami plot}
\usage{
prep_miami_data(data, split_by, split_at, chr = "chr", pos = "pos", p = "p")
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{split_by}{A character vector. The name of the column to use for
splitting into top and bottom sections of the Miami plot. Required.}

\item{split_at}{A character or numeric vector. If numeric, the top plot will
contain your results where the values in the \code{split_by} column are
>= \code{split_at}. If character, top plot will contain your results where
the values in the \code{split_by} column are equal to \code{split_at}.
Required.}

\item{chr}{The name of the column containing your chromosome information.
Defaults to "chr"}

\item{pos}{The name of the column containing your position information.
Defaults to "pos"}

\item{p}{The name of the column containing your p-value information.
Defaults to "p"}
}
\value{
A list containing the data needed for the top and bottom plots,
  axes, and the maximum p-value (for sizing the plot)
}
\description{
Prepare data for miami plot
}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
