% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miami_input.R
\name{check_miami_input}
\alias{check_miami_input}
\title{Check input to ggmiami function}
\usage{
check_miami_input(data, split.by, split.at)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{split.by}{A character vector. The name of the column to use for
splitting into top and bottom sections of the Miami plot. Required.}

\item{split.at}{A character or numeric vector. If numeric, the top plot will
contain your results where the values in the \code{split.by} column are
>= \code{split.at}. If character, top plot will contain your results where
the values in the \code{split.by} column are equal to \code{split.at}. Required.}
}
\value{
Will return an error message if there are problems.
}
\description{
Check input to ggmiami function
}
\examples{
# If you would like to put positive beta values in the top of your plot, and
# negative values on the bottom.
check_miami_input(data = df, split.by = "Beta", split.at = 0)

# If you want genotyped SNPs on top, and imputed SNPs on the bottom plot.
check_miami_input(data = df, split.by = "GenoStatus", split.at = "Genotyped")

# If you want to compare results from two studies, with study A on top.
check_miami_input(data = df, split.by = "Study", split.at = "A")

}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
