% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmiami.R
\name{ggmiami}
\alias{ggmiami}
\title{Create a Miami plot ggplot2 object.}
\usage{
ggmiami(
  data,
  split.by,
  split.at,
  chr = "chr",
  pos = "pos",
  p = "p",
  genomewideline = 5e-08,
  suggestiveline = 1e-05,
  hits.label = NULL,
  top.n.hits = 5
)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{split.by}{A character vector. The name of the column to use for
splitting into top and bottom sections of the Miami plot. Required.}

\item{split.at}{A character or numeric vector. If numeric, the top plot will
contain your results where the values in the \code{split.by} column are
>= \code{split.at}. If character, top plot will contain your results where
the values in the \code{split.by} column are equal to \code{split.at}. Required.}

\item{chr}{The name of the column containing your chromosome information.
Defaults to "chr"}

\item{pos}{The name of the column containing your position information.
Defaults to "pos"}

\item{p}{The name of the column containing your p-value information.
Defaults to "p"}

\item{genomewideline}{Should we draw a genome-wide significance line? Set to
NULL if you do not want this line. Defaults to 5e-8, or supply your own number.}

\item{suggestiveline}{Should we draw a suggestive significance line? Set to
NULL if you do not want this line. Defaults to 1e-5, or supply your own number.}

\item{hits.label}{Either the name of the column(s), max. 2, to use for
automatically labeling n hits, determined using \code{top.n.hits}, or a
character vector of probes/genes/SNPs to label. If supplying a list of
genes, it is helpful to also supply \code{top.n.hits} to limit the labels
to the top N results, though this isn't necessary.
Defaults to NULL: labels aren't displayed.}

\item{top.n.hits}{How many of the top hits do you want to label? Defaults to
5. Set to NULL to turn off this filtering.}
}
\value{
a \code{ggplot2} object
}
\description{
Create a Miami plot ggplot2 object.
}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
\keyword{SNP}
\keyword{genetics}
\keyword{manhattan}
\keyword{miami}
\keyword{plot}
