% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmiami.R
\name{ggmiami}
\alias{ggmiami}
\title{Create a Miami plot ggplot2 object.}
\usage{
ggmiami(
  data,
  split_by,
  split_at,
  chr = "chr",
  pos = "pos",
  p = "p",
  chr_colors = c("black", "grey"),
  top_ylab = "-log10(p)",
  bottom_ylab = "-log10(p)",
  genome_line = 5e-08,
  suggestive_line = 1e-05,
  hits_label_col = NULL,
  hits_label = NULL,
  top_n_hits = 5
)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{split_by}{A character vector. The name of the column to use for
splitting into top and bottom sections of the Miami plot. Required.}

\item{split_at}{A character or numeric vector. If numeric, the top plot will
contain your results where the values in the \code{split_by} column are
>= \code{split_at}. If character, top plot will contain your results where
the values in the \code{split_by} column are equal to \code{split_at}.
Required.}

\item{chr}{The name of the column containing your chromosome information.
Defaults to "chr"}

\item{pos}{The name of the column containing your position information.
Defaults to "pos"}

\item{p}{The name of the column containing your p-value information.
Defaults to "p"}

\item{chr_colors}{Either a vector of two colors to alternate across
chromosomes or a vector of colors to use for coloring chromosomes, with
length equal to the number of chromosomes being plotted. Defaults to
alternating black and grey.}

\item{top_ylab}{What would you like the y-axis title for the top plot to be?
Defaults to "-log10(p)". Text added here will result in a two-line axis
with your text on top and "-log10(p)" below.}

\item{bottom_ylab}{Same as \code{top_ylab}, but for the bottom plot.}

\item{genome_line}{Should we draw a genome-wide significance line? Set to
NULL if you do not want this line. Defaults to 5e-8, or supply your own
number.}

\item{suggestive_line}{Should we draw a suggestive significance line? Set to
NULL if you do not want this line. Defaults to 1e-5, or supply your own
number.}

\item{hits_label}{A user-specified character vector of probes/genes/SNPs
to label. Defaults to NULL.}

\item{top_n_hits}{How many of the top hits do you want to label? Defaults to
5. Set to NULL to turn off this filtering.}

\item{hits_labe_lcol}{Either the name of the column(s), max = 2, to use for
automatically labeling n hits, determined using \code{top_n_hits}, or the
column where the values you provide in \code{hits_label} can be found.
Defaults to NULL: labels aren't displayed.}
}
\value{
a \code{ggplot2} object
}
\description{
Create a Miami plot ggplot2 object.
}
\examples{
  If you want to put SNPs with positive beta values on top and netative beta
  values on the bottom:
  ggmiami(data = df, split_by = "beta", split_at = 0)

  If you want to put results from study A on top and study B on bottom:
  ggmiami(data = df, split_by = "study", split_at = "A")

  If you want to add a genome wide line at 9e-8, instead of 5e-8:
  ggmiami(data = df, split_by = "study", split_at = "A", genome_line = 9e-8)

  If you want to label the top 5 hits with the SNP and gene name:
  ggmiami(data = df, split_by = "study", split_at = "A",
          hits_label_col = c("SNP", "Gene"))

  If you want to label the top 10 hits from the AHRR, TBX15, and RARA genes:
  ggmiami(data = df, split_by = "study", split_at = "A",
          hits_label_col = "Gene", hits_label = c("AHRR", "TBX15", "RARA"),
          top_n_hits = 10)

  When labeling, the items in hits_label must all come from a single column.
  Specifying multiple columns will return an error.

  \dontrun{
  ggmiami(data = df, split_by = "study", split_at = "A",
          hits_label_col = c("chr", Gene"),
          hits_label = c("1", AHRR", "2", "TBX15", "RARA"),
          top_n_hits = 10)
  }

}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
\keyword{SNP}
\keyword{genetics}
\keyword{manhattan}
\keyword{miami}
\keyword{plot}
