% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_miami_labels.R
\name{make_miami_labels}
\alias{make_miami_labels}
\title{Make label data.frame for miami plot}
\usage{
make_miami_labels(data, p = "loggedp", hits.label, top.n.hits = 5)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{p}{The name of the column containing your logged p-value information.
Defaults to "loggedp," assuming data preparation with `prep_miami_data`}

\item{hits.label}{Either the name of the column(s), max. 2, to use for
automatically labeling n hits, determined using \code{top.n.hits}, or a
character vector of probes/genes/SNPs to label. If supplying a list of
genes, it is helpful to also supply \code{top.n.hits} to limit the labels
to the top N results, though this isn't necessary.}

\item{top.n.hits}{How many of the top hits do you want to label? Defaults to
5. Set to NULL to turn off this filtering, but this is not recommended.}
}
\value{
A data.frame with
}
\description{
Make label data.frame for miami plot
}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
