% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_miami_labels.R
\name{make_miami_labels}
\alias{make_miami_labels}
\title{Make label data.frame for miami plot}
\usage{
make_miami_labels(
  data,
  hits.label.col,
  hits.label = NULL,
  top.n.hits = 5,
  loggedp = "loggedp"
)
}
\arguments{
\item{data}{A data.frame object. Required.}

\item{hits.label.col}{Either the name of the column(s), max. 2, to use for
automatically labeling n hits, determined using \code{top.n.hits}, or the
column where the values you provide in \code{hits.label} can be found. Required.}

\item{hits.label}{A user-specified character vector of probes/genes/SNPs
to label. Defaults to NULL.}

\item{top.n.hits}{How many of the top hits do you want to label? Defaults to
5. Set to NULL to turn off this filtering, but this is not recommended
because the plot can get cluttered easily.}

\item{loggedp}{The name of the column containing your logged p-value information.
Defaults to "loggedp," assuming data preparation with `prep_miami_data`}
}
\value{
A data.frame with
}
\description{
Make label data.frame for miami plot
}
\references{
\url{https://github.com/juliedwhite/miamiplot}
}
\author{
Julie White
}
