% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamiplot.R
\docType{package}
\name{miamiplot-package}
\alias{miamiplot}
\alias{miamiplot-package}
\title{miamiplot: Create a ggplot2 miami plot}
\description{
This package can be used to create a Miami plot, with the 
    chromosome labels bisecting the middle of the plot. The output is a ggplot2
    object. Users can specify which data they want plotted on top vs. bottom
    , whether to display significance line(s), what colors to give chromosomes,
    and what points to label.
}
\author{
\strong{Maintainer}: Julie White \email{julie.white@nih.gov}

}
\keyword{internal}
